pyproject.toml
[tool.poetry]
name = "tortoise-orm-stubs"
version = "1.0.0"
description = "Type stubs that make tortoise-orm a lot easier to work with when using type checkers."
packages = [{ include = "tortoise-stubs" }]
authors = ["Stanislav Zmiev <szmiev2000@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Ovsyanka83/tortoise-orm-stubs"

[tool.poetry.dependencies]
python = "^3.8"
tortoise-orm = "*"

[tool.poetry.dev-dependencies]
pyupgrade = "^2.37.3"
black = "^22.8.0"
autoflake = "^1.5.3"
pytest = "^7.1.3"
pytest-cov = "^3.0.0"
isort = "^5.10.1"
mypy = "^0.971"
pysh = "^3.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.html]
skip_covered = true
skip_empty = true

[tool.coverage.report]
omit = []
fail_under = 90
skip_covered = true
skip_empty = true
show_missing = true

[tool.isort]
profile = "black"
multi_line_output = 3
skip_gitignore = true
skip_glob = []
line_length = 120
lines_between_sections = 1

[tool.black]
line-length = 120
target_version = ['py38']
